Check in test data and add scripts for CI
