+++ .gitignore
@@ -47,5 +47,4 @@ amplifytools.xcconfig
 
 out/
 *debug.log
-firebase-exports
 *.tsbuildinfo
