+++ integrationTests/exportedTestData/auth_export/accounts.json
@@ -1,0 +1,1 @@
+{"kind":"identitytoolkit#DownloadAccountResponse","users":[{"localId":"OrSh0WyGg8bWsGETHMKgy2hroV7W","createdAt":"1644867431060","lastLoginAt":"1644867431062","displayName":"Panda Raccoon","providerUserInfo":[{"providerId":"google.com","rawId":"6987557938223822050477752443492343737557","federatedId":"6987557938223822050477752443492343737557","displayName":"Panda Raccoon","email":"panda.raccoon.439@example.com","screenName":"raccoon_panda"}],"validSince":"1644955192","email":"panda.raccoon.439@example.com","emailVerified":true,"disabled":false},{"localId":"yrkjhff1GFTIHbUVL7kaNZnzdwXp","createdAt":"1644958440415","lastLoginAt":"1644958440414","emailVerified":false,"email":"test@example.com","salt":"fakeSalt5VyKns8rG57x4TD87FoB","passwordHash":"fakeHash:salt=fakeSalt5VyKns8rG57x4TD87FoB:password=password","passwordUpdatedAt":1644958440415,"validSince":"1644958440","providerUserInfo":[{"providerId":"password","email":"test@example.com","federatedId":"test@example.com","rawId":"test@example.com","displayName":"Conan O&#39;Brien"}],"lastRefreshAt":"2022-02-15T20:54:00.415Z","displayName":"Conan O&#39;Brien"}]}
