+++ integrationTests/exportedTestData/auth_export/config.json
@@ -1,0 +1,1 @@
+{"signIn":{"allowDuplicateEmails":false},"usageMode":"DEFAULT"}
