+++ integrationTests/exportedTestData/firebase-export-metadata.json
@@ -1,0 +1,12 @@
+{
+  "version": "10.1.2",
+  "firestore": {
+    "version": "1.13.1",
+    "path": "firestore_export",
+    "metadata_file": "firestore_export/firestore_export.overall_export_metadata"
+  },
+  "auth": {
+    "version": "10.1.2",
+    "path": "auth_export"
+  }
+}
