+++ scripts/test-integration-ci.js
@@ -1,0 +1,17 @@
+const { demoProjectId, env, runOrExit } = require("./common")
+
+runOrExit("yarn", ["--cwd", "functions", "build"], { stdio: "inherit" })
+runOrExit(
+  "firebase",
+  [
+    "--project",
+    demoProjectId,
+    "emulators:exec",
+    "--only",
+    "auth,functions,pubsub,firestore",
+    "--import",
+    "integrationTests/exportedTestData",
+    "yarn test:integration --forceExit"
+  ],
+  { stdio: "inherit", env }
+)
